i8 quantised, bench: 8234605

+++ src/policy.h
@@ -29,7 +29,7 @@ constexpr float QB = 128;
 constexpr float QAB = QA * QB;
 
 struct PolicyNetwork {
-    alignas(32) std::array<int16_t, p_l1Size * p_inputSize> featureWeights;
+    alignas(32) std::array<int8_t, p_l1Size * p_inputSize> featureWeights;
     alignas(32) std::array<int16_t, p_l1Size> featureBiases;
     alignas(32) std::array<int16_t, p_outputCount * p_l1Size> outputWeights;
     alignas(32) std::array<int, p_outputCount> outputBiases;
